Move topsort to module traits
